+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+
 # Changelog
 
 This changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- Possible headings: Added, Changed, Deprecated, Fixed, Removed, Security -->
+
+## Unreleased
+
+### Added
+
+- auto-paged help output
 
 ## [0.21.0] - 2026-06-12